requirements.txt
streamlit
python-dotenv
langchain
langchain_community
groq
langchain_groq